pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = ["Valera <Valera@valera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
coverage = "^7.2.7"
django-filter = "^23.2"
social-auth-app-django = "^5.2.0"
django-debug-toolbar = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
